Fixed portal css for small screens and Firefox.
Firefox needs an explicit declaration to hide the scrollbar when
expanding the schedule on a small screen. Also, because of some
earlier unknown change, the schedule had an x overflow on small
screens because the text would not wrap. I set a minimum width for
the schedule, but I also set more compressed schedule css for more
narrow screens.